+++ src/components/items/techListItem.tsx
